refactor(eclipse): move regex for leading tabs to stringUtils.
