+++ clients/eclipse/plugin/src/com/tabbyml/tabby4eclipse/StringUtils.java
@@ -1,9 +1,11 @@
 package com.tabbyml.tabby4eclipse;
 
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
 import org.eclipse.swt.graphics.RGB;
 
 public class StringUtils {
-
 	public static String escapeCharacters(String jsonString) {
 		return jsonString.replace("\\", "\\\\").replace("\"", "\\\"").replace("\n", "\\n").replace("\r", "\\r")
 				.replace("\t", "\\t");
@@ -36,5 +38,34 @@ public class StringUtils {
 		s *= 100;
 		l *= 100;
 		return String.format("%.0f, %.0f%%, %.0f%%", h, s, l);
+	}
+
+	public static class TextWithTabs {
+		private int tabs;
+		private String text;
+
+		public TextWithTabs(int tabs, String text) {
+			this.tabs = tabs;
+			this.text = text;
+		}
+
+		public int getTabs() {
+			return tabs;
+		}
+
+		public String getText() {
+			return text;
+		}
+	}
+
+	static final Pattern PATTERN_LEADING_TABS = Pattern.compile("^(\\t*)(.*)$");
+
+	public static TextWithTabs splitLeadingTabs(String text) {
+		Matcher matcher = PATTERN_LEADING_TABS.matcher(text);
+		if (matcher.matches()) {
+			return new TextWithTabs(matcher.group(1).length(), matcher.group(2));
+		} else {
+			return new TextWithTabs(0, text);
+		}
 	}
 }
